Add files via upload
Dijkstra's Algorithm Using Fibonacci Heap in C:
Dijkstra's original shortest path algorithm does not use a priority queue and runs in O(V2) time. When using a Fibonacci heap as a priority queue, it runs in O(E + V log V) time, which is asymptotically the fastest known time complexity for this problem.